Create recurseFileToTrash_Except
Program to read a directory(say parent_dir) and read each and every sub-directory(with 3-letter name eg..123, 456,..) in the given directory. You are given a list of file names without extensions. If the sub directories are having pdf-files with name present in the list, then delete the file and move to trash/recycle bin. Else move the file with respective sub-directory to a new directory, named "exception".