Create recurseFileToTrash_Except_PathFromXML
Python script to read the path of a directory from any XML file, read each and every sub-directory of the given directory. You are given a list of file names without extensions. If the sub directories are having files with name present in the list, then delete the file and move to trash/recycle bin. Else move the file with respective sub-directory to a Path as mentioned in the XML file as destination.